lambdas/image-processor/pyproject.toml
[project]
name = "image-processor"
version = "0.1.0"
description = ""
authors = [
    {name = "Jose Suarez",email = "jose95sp@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "boto3 (>=1.40.41,<2.0.0)"
]

[tool.poetry]
packages = [{include = "image_processor", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
